Run strange partial frame when button class subclasses SimpleButton
It seems that we don't do a full `run_inner_goto_frame` after all -
we do *not* run any `construct_frame` calls. This results in
constructors for other children not being run in this nested frame,
so framescripts will only see sibiling of the button initialized
if the sibling comes before the button.